[2607-DEV-676] fix: one shared Supabase write-target guard for the seed scripts
The substring guard `url.includes(DEV_PROJECT_REF)` was live in four seed
scripts, not the two first reported: seed-smoke-guide.js,
seed-clerk-test-users.js and seed-guest-test-user.js all still accepted
https://evil.example/?ref=<ref>, https://<ref>.supabase.co.evil.example
and plaintext http to the hosted project — then sent a service-role key.

Four private copies of one security check are why the same defect had to
be fixed twice on this PR and stayed live in the siblings. Extracted to
scripts/lib/safe-supabase-target.js (matching the existing scripts/lib/
convention) so there is one place to harden.

Probed 6 URLs x 4 scripts, 24/24 correct: plaintext-hosted,
substring-in-query, suffixed-hostname and the PROD ref all refused;
HTTPS-DEV and localhost-http accepted.

check-env.js keeps its includes() deliberately — it is an advisory
classifier printing which project you point at, not a write guard.
